begin creating a syntax tree

+++ include/tokenizer.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <iostream>
 #include <optional>
 #include <string>
 #include <vector>
@@ -21,7 +22,7 @@ private:
   [[nodiscard]] std::optional<char> peak(int ahead = 1) const;
   char consume();
 
-  int m_index = 0;
+  size_t m_index = 0;
   const std::string m_src;
 };
 } // namespace Tokenizer_NS
